requirements.txt
pandas==2.2.2
numpy==1.26.4
yfinance==0.2.40
ta==0.11.0
scikit-learn==1.5.1
xgboost==2.1.1
tensorflow==2.16.1
fastapi==0.115.0
uvicorn==0.30.1
streamlit==1.38.0
plotly==5.22.0
evidently==0.4.32
dvc==3.58.0
apscheduler==3.10.4
shap==0.48.0
pyarrow==15.0.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
mlflow==2.15.1